+++ crates/common/src/ens.rs
@@ -16,13 +16,6 @@ sol! {
     contract EnsRegistry {
         /// Returns the resolver for the specified node.
         function resolver(bytes32 node) view returns (address);
-        function setSubnodeRecord(
-                bytes32 node,
-                bytes32 label,
-                address owner,
-                address resolver,
-                uint64 ttl
-            ) external;
     }
 
     /// ENS Resolver interface.
@@ -33,33 +26,6 @@ sol! {
 
         /// Returns the name associated with an ENS node, for reverse records.
         function name(bytes32 node) view returns (string);
-    }
-
-    #[sol(rpc)]
-    contract NameWrapper {
-        function isWrapped(bytes32 node) external returns (bool);
-        function setSubnodeRecord(
-                bytes32 node,
-                string label,
-                address owner,
-                address resolver,
-                uint64 ttl,
-                uint32 fuses,
-                uint64 expiry,
-            ) external;
-    }
-
-    #[sol(rpc)]
-    contract PublicResolver {
-        function setAddr(bytes32 node, address addr) external;
-        function addr(bytes32 node) external returns (address);
-        function setName(bytes32 node,  string newName) external;
-    }
-
-    #[sol(rpc)]
-    contract ReverseRegistrar {
-        function setName(string memory name) external returns (bytes32);
-        function setNameForAddr(address addr, address owner, address resolver, string name) external;
     }
 }
