+++ crates/forge/Cargo.toml
@@ -27,6 +27,7 @@ foundry-evm.workspace = true
 foundry-wallets.workspace = true
 foundry-linking.workspace = true
 forge-script-sequence.workspace = true
+naming.workspace = true
 
 comfy-table.workspace = true
 eyre.workspace = true
